Added in some config file support and change aware actions.
Uses $XDG_CONFIG_HOME/til, $HOME/.tilrc, /etc/tilrc in that order to find
configuration variables, which are currently just `TILHOME` and `DOC_SUFFIX`.

Added return codes and state checks to dictate whether the file is created and
added to the git repository or not.  Also "generisized" the editor interface in
preparations of allowing other editor support.